create commonResponses file and moved 501 response there
